app/requirements.txt
hopsworks==3.2.0
hsfs==3.2.1
matplotlib==3.7.2
numpy==1.24.4
pandas==1.3.5
pendulum==2.1.2
# psycopg2==2.9.5
psycopg2_binary==2.9.7
PyYAML==6.0.1
redis==4.6.0
streamlit==1.27.0
torch==2.0.1
